Switch license, update README and dependencies

+++ src/mouse.rs
@@ -69,7 +69,7 @@ impl<'c> Mouse<'c> {
         self.controller.write_mouse(command as u8)?;
         self.check_response()?;
         if let Some(data) = data {
-            self.controller.write_data(data as u8)?;
+            self.controller.write_data(data)?;
             self.check_response()?;
         }
         Ok(())
